+++ src/shared/helpers/index.ts
@@ -1,3 +1,3 @@
 export * from './generate-id';
 export * from './get-host';
-export * from './mutate-object-properties';
+export * from './objects';
